+++ cli/create/builtin_templates/builtin_templates.go
@@ -11,7 +11,6 @@ var TemplatesFs embed.FS
 
 // FileModes contains mapping of file modes by built-in template name.
 var FileModes = map[string]map[string]int{
-	"cartridge":       static.CartridgeFileModes,
 	"vshard_cluster":  static.VshardClusterFileModes,
 	"single_instance": static.SingleInstanceFileModes,
 	"config_storage":  static.ConfigStorageFileModes,
@@ -20,7 +19,6 @@ var FileModes = map[string]map[string]int{
 
 // Names contains built-in template names.
 var Names = [...]string{
-	"cartridge",
 	"vshard_cluster",
 	"single_instance",
 	"config_storage",